src/pyproject.toml
[tool.poetry]
name = "smit-studio"
version = "0.1.0"
description = ""
authors = ["Roman Pavlyuchenko <romanpavlyuchenko@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
aerich = "^0.7.1"
tortoise-orm = "^0.19.3"
gunicorn = "^20.1.0"
uvicorn = "^0.22.0"
asyncpg = "^0.28.0"
flake8 = "^6.0.0"
python-multipart = "^0.0.6"

[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "migrations/"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
